rekor: refactor service wait init containers
The goal of this change is to upgrade curl to the current stable release. However, since the current stable release of curl behaves differently with respect to non-HTTP servers, a different solution was necessary for determining availability of the database (MySQL and MariaDB do not use HTTP).

This resulted in three changes:

1. Change the curl image to docker.io/curlimages/curl:7.79.1
2. Modify the test for database server availability to use netcat instead of curl
3. Refactor the initContainerImage section of the values.yaml to support multiple init containers

Signed-off-by: Gregory Moyer <gregory.moyer@nationwide.com>