move default into template helper to allow overwrite.
Due to the nature of helm handling of maps, it is not possible to
overwrite keys in the base value.yaml of the chart. Keys from value.yaml
included with the chart will always be present.

If it was the main chart, it would be possible to null existing keys,
but due to https://github.com/helm/helm/issues/9136, it does not work
with sub-charts at the moment.

So to allow downstream users to overwrite the config contents, the
default is now set in the template helper, when an empty object is
specified.

Signed-off-by: Kenny Leung <kleung@chainguard.dev>