Add rekor-createtree helm chart for creating a new tree for rekor
This will be used for sharding. Initially, I thought we could just force the `createtree` Job through the rekor Chart.
But, it looks like that will automatically restart the rekor server, which might lead to a race between updating the Rekor configmap with the new tree ID and starting up the server.

With a separate Chart, we can create a new tree and no effect anything else.

Signed-off-by: Priya Wadhwa <priya@chainguard.dev>