Expose the deployment strategy values for the policy controller
Prior to this change, the policy controller webhook was not able to have
its deployment strategy modified. If you only deployed a single replica,
it could not perform a rolling update due to the default `maxSurge:
25%` being rounded down to 0.

This change exposes those values, so that the `maxSurge` can be updated
and a single instance can be rolled.

Fixes #748.

Signed-off-by: Alex Shearn <alex.shearn@kraken.tech>