rekor: Added support for AWS KMS settings.
Prior to this commit, the chart did not provide a way to supply
AWS credentials for AWS KMS. This commit adds support for AWS KMS
by allowing users to supply an AWS region ID and IAM credentials.

AWS KMS users must specify the "kmsType" parameter with a value
of "aws". The chart will then look for a kubernetes secret named
by the "awsKmsCredentialsSecret" parameter.

The AWS region ID can be supplied using the newly-added
"awsKmsRegion" parameter.

Signed-off-by: Stephen Fox <sfoxjr@equinix.com>