Remove rekor-tiles Ingress in favor of NEGs
The rekor-tiles service does not directly enable its own read path, but
instead relies on external routing to a GCS bucket. The GCP Load
Balancer created by a Kubernetes Ingress resource is not capable of
routing to anything outside of Kubernetes, so this load balancer must be
managed in Terraform instead.

This change removes the rekor-tiles Ingress and supporting resources
entirely, and instead adds a Network Endpoint Group annotation to the
Service resource, which will create the Network Endpoint Group[1] in
GCP, which Terraform can then use to route to the cluster.

[1] https://cloud.google.com/load-balancing/docs/negs/zonal-neg-concepts

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>