Conditionally set Rekor Redis argument flags (#1009)
This changes the logic for how redis.hostname is set. Either Redis is
deployed via K8s, at which point the hostname is either set to the
fullname by default or the provided hostname, or Redis is externally
managed at which point the hostname is set to the provided hostname. The
flags to Rekor will only be set when hostname is not empty, so they will
not be set if Redis is disabled or an empty hostname is provided, which
is the default.

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>
Co-authored-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>
Signed-off-by: Philippe Arcand <7255842+philarcand@users.noreply.github.com>