Add graceful shutdown settings for rekor-tiles (#1022)
Add a preStop hook to delay terminating the pod for 15 seconds while the
load balancer drains the endpoint's connections. The delay matches the
drain timeout set on the load balancer.

Set the terminationGracePeriodSeconds to 65 seconds, which is:
    preStop hook (15s)
  + default server idle timeout set in rekor-tiles binary (20s)
  + default log connection timeout set in rekor-tiles binary (30s)

See https://victoriametrics.com/blog/go-graceful-shutdown/index.html for
information about this recommendation.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>