OTel attributes and termination grace period fix (#1026)
* Add OTEL_RESOURCE_ATTRIBUTES for rekor-tiles

Set OTEL_RESOURCE_ATTRIBUTES to populate service_instance_id and
service_namespace in OpenTelemetry metrics.

This will make it possible to distinguish metrics from different
instances and hopefully reduce or eliminate "TimeSeries could not be
written...Points must be written in order. One or more of the points
specified had an older start time than the most recent point." errors.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>

* Fix termination grace period field

This is a pod field, not a container field.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>

---------

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>