Add ctlog-tiles chart (#1098)
Add a chart named ctlog-tiles to support TesseraCT as an alternative CT
log for Fulcio. The POSIX and GCP backends are supported.

This chart has a unique way of crafting the image string based on which
personality (posix or gcp) is chosen. This lets values.yaml/Chart.yaml
still be the source of truth for the image version, which is consistent
with how Sigstore charts are used in the public good instance (as
opposed to overriding the image version in the infrastructure
values.yaml).

This is closely based on the rekor-tiles chart. It creates NEGs but no
explicit Ingress gateway, since a multi-purpose load balancer needs to
be created externally to link the service to the backend bucket.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>